refactor: extract backup as single components
